+++ backend/src/api/compensator/content-types/compensator/schema.json
@@ -44,14 +44,18 @@
       "type": "richtext"
     },
     "compensationProgressImage": {
+      "type": "media",
+      "multiple": false,
+      "required": false,
       "allowedTypes": [
         "images",
         "files",
         "videos",
         "audios"
-      ],
-      "type": "media",
-      "multiple": false
+      ]
+    },
+    "compensationProgressTitle": {
+      "type": "string"
     }
   }
 }
